+++ CHANGELOG.md
@@ -29,6 +29,7 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 - Gate: Removed constraint checks when upserting business partner relations ([1288](https://github.com/eclipse-tractusx/bpdm/issues/1288))
 - App: Renamed api attributes/fields with corresponding 25.06 standards for BPDM Pool, Gate and Orchestrator Services([#1304](https://github.com/eclipse-tractusx/bpdm/issues/1304))
 - Gate: Changed endpoint paths to better differentiate between business partner or relation endpoints ([#1327](https://github.com/eclipse-tractusx/bpdm/issues/1327))
+- Pool: Enforce dataspace participant validation for managing entity in IsManagedBy relation([#1390](https://github.com/eclipse-tractusx/bpdm/issues/1390))
 
 ### Known Knowns